fix: add local aliases to webpack.dev-stage.config.js
