+++ config/webpack.dev.config.js
@@ -4,7 +4,6 @@
 const { merge } = require('webpack-merge');
 const Dotenv = require('dotenv-webpack');
 const dotenv = require('dotenv');
-const fs = require('fs');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
 const PostCssAutoprefixerPlugin = require('autoprefixer');
@@ -15,6 +14,7 @@ const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin'
 const commonConfig = require('./webpack.common.config.js');
 const presets = require('../lib/presets');
 const resolvePrivateEnvConfig = require('../lib/resolvePrivateEnvConfig');
+const getLocalAliases = require('./getLocalAliases');
 
 // Add process env vars. Currently used only for setting the
 // server port and the publicPath
@@ -27,88 +27,14 @@ dotenv.config({
 // in temporary modifications to .env.development.
 resolvePrivateEnvConfig('.env.private');
 
-/*
-This function reads in a 'module.config.js' file if it exists and uses its contents to define
-a set of webpack resolve.alias aliases for doing local development of application dependencies.
-It reads the package.json file of the dependency to determine if it has any peer dependencies, and
-then forces those peer dependencies to be resolved with the application's version.  Primarily, this
-is useful for making sure there's only one version of those dependencies loaded at once, which is a
-problem with both react and react-intl.
-
-The module.config.js file should have the form:
-
-{
-  localModules: [
-    { moduleName: 'nameOfPackage', dir: '../path/to/repo', dist: '/path/to/dist/in/repo' },
-    ... others...
-  ],
-}
-
-Some working examples, as of the time of this writing:
-
-{ moduleName: '@edx/paragon/scss', dir: '../paragon', dist: 'scss' }
-{ moduleName: '@edx/paragon', dir: '../paragon', dist: 'dist' }
-{ moduleName: '@edx/frontend-platform', dir: '../frontend-platform', dist: 'dist' }
-
-*/
-function getLocalAliases() {
-  const aliases = {};
-
-  try {
-    const moduleConfigPath = path.resolve(process.cwd(), 'module.config.js');
-    if (!fs.existsSync(moduleConfigPath)) {
-      console.log('No local module configuration file found. This is fine.');
-      return aliases;
-    }
-    // eslint-disable-next-line import/no-dynamic-require, global-require
-    const { localModules } = require(moduleConfigPath);
-
-    let allPeerDependencies = [];
-    const excludedPeerPackages = [];
-    if (localModules.length > 0) {
-      console.info('Resolving modules from local directories via module.config.js.');
-    }
-    localModules.forEach(({ moduleName, dir, dist = '' }) => {
-      console.info(`Using local version of ${moduleName} from ${dir}/${dist}.`);
-      // eslint-disable-next-line import/no-dynamic-require, global-require
-      const { peerDependencies = {}, name } = require(path.resolve(process.cwd(), dir, 'package.json'));
-      allPeerDependencies = allPeerDependencies.concat(Object.keys(peerDependencies));
-      aliases[moduleName] = path.resolve(process.cwd(), dir, dist);
-      excludedPeerPackages.push(name);
-    });
-
-    allPeerDependencies = allPeerDependencies.filter((dep) => !excludedPeerPackages.includes(dep));
-
-    allPeerDependencies.forEach((dep) => {
-      aliases[dep] = path.resolve(process.cwd(), 'node_modules', dep);
-    });
-  } catch (e) {
-    console.error(e);
-    console.error('Error in module.config.js parsing. module.config.js will be ignored.');
-    return {};
-  }
-  return aliases;
-}
-
 const aliases = getLocalAliases();
 const PUBLIC_PATH = process.env.PUBLIC_PATH || '/';
 
 module.exports = merge(commonConfig, {
   mode: 'development',
   devtool: 'eval-source-map',
-  entry: {
-    // enable react's custom hot dev client so we get errors reported in the browser
-    hot: require.resolve('react-dev-utils/webpackHotDevClient'),
-    app: path.resolve(process.cwd(), 'src/index'),
-  },
   output: {
     publicPath: PUBLIC_PATH,
-  },
-  optimization: {
-    // ``runtimeChunk`` is needed to support hot reloading. See
-    // https://github.com/pmmmwh/react-refresh-webpack-plugin/issues/88#issuecomment-627558799
-    // for more details.
-    runtimeChunk: 'single',
   },
   resolve: {
     alias: aliases,
